// src/lib/api/fs/v1.d.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */


export interface paths {
  "/fileTree": {
    /** List paths in root */
    get: {
      responses: {
        /** @description Success */
        200: {
          content: {
            "application/json": components["schemas"]["FileTreeEntry"][];
          };
        };
      };
    };
    /** Add new path */
    post: {
      requestBody?: {
        content: {
          "application/json": components["schemas"]["FileTreeEntry"];
        };
      };
      responses: {
        /** @description Path created */
        201: {
          content: {
            "application/json": components["schemas"]["FileTreePath"];
          };
        };
        400: components["responses"]["NonLocalPath"];
        /** @description Path already exists */
        409: {
          content: {
            "application/json": components["schemas"]["Error"];
          };
        };
      };
    };
  };
  "/fileTree/{path}": {
    /** List nested paths */
    get: {
      parameters: {
        path: {
          path: string;
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: {
            "application/json": components["schemas"]["FileTreeEntry"][];
          };
        };
        400: components["responses"]["NonLocalPath"];
        404: components["responses"]["PathNotFound"];
      };
    };
    /** Delete path */
    delete: {
      parameters: {
        path: {
          path: string;
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: {
            "application/json": components["schemas"]["FileTreePath"];
          };
        };
        400: components["responses"]["NonLocalPath"];
        404: components["responses"]["PathNotFound"];
      };
    };
    parameters: {
      path: {
        path: string;
      };
    };
  };
  "/file/{filePath}": {
    /** Download plain/text file */
    get: {
      parameters: {
        path: {
          filePath: string;
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: {
            "plain/text": string;
          };
        };
        400: components["responses"]["BadRequest"];
        404: components["responses"]["PathNotFound"];
      };
    };
    /** Update file */
    put: {
      parameters: {
        path: {
          filePath: string;
        };
      };
      requestBody?: {
        content: {
          "plain/text": string;
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: never;
        };
        400: components["responses"]["BadRequest"];
        404: components["responses"]["PathNotFound"];
      };
    };
  };
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
    FileTreePath: {
      /** @example /path/to/file */
      path: string;
    };
    FileTreeEntry: {
      /** @example /path/to/file */
      path: string;
      isDir: boolean;
    };
    Error: {
      message: string;
    };
  };
  responses: {
    /** @description Non local path */
    NonLocalPath: {
      content: {
        "application/json": components["schemas"]["Error"];
      };
    };
    /** @description Path not found */
    PathNotFound: {
      content: {
        "application/json": components["schemas"]["Error"];
      };
    };
    /** @description Bad request ("Non local path." or "Not plain/text file.") */
    BadRequest: {
      content: {
        "application/json": components["schemas"]["Error"];
      };
    };
  };
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}

export type $defs = Record<string, never>;

export type external = Record<string, never>;

export type operations = Record<string, never>;